📚 Improved script to generate device documentation.
Removed the dev dependency on zigbee-herdsman-converters as this wasn't
always working correctly for me. The newly added bash script will figure
out what version of zigbee-herdsman-converters is used by the latest
release of Zigbee2MQTT, install that (without adding it to the
package.json) and then run the docgen script.